+++ CHANGELOG.md
@@ -18,6 +18,8 @@ We use the following categories for changes:
 
 ### Added
 
+- kube state metrics flag [#sc-8036]
+
 ### Changed
 
 ### Fixed
@@ -31,7 +33,9 @@ We use the following categories for changes:
 ## [0.8.7] 2023-03-27
 
 ### Added
+
 - custom metrics flag [#sc-7824]
+
 ### Changed
 
 ### Fixed